Merge feature branch for new simulator implementation (#23)
The goal of the new version is to be able to simulate possible player
moves, and track probability of each resulting state. The goal is to
find any riskless path through a battle.

The huge diff is mostly the result of some build system changes made
along the way:

1. Drop ts from dist -- it was a mistake that these were there to begin
with
2. Add npm commands for build, watch, serve
3. Configure VS Code to have the Jest extension for running tests